fixes #4 --- removed closeMobileMenu call from desktop condition in setupMobileMenu, as the layout shift was required to run any time the desktop view loads or the viewport shifts from mobile to desktop.
If the mobile menu was left open when the breakpoint to desktop is crossed, the mobile menu is not visibly open in desktop mode but remained open if you switch back to mobile view. To prevent this, I added closeMobileMenu() to the mobile breakpoint condition instead. because closeMobileMenu already sets inert on topnavMenu, I also removed the set inert instruction in the mobile condition of setupMobileMenu.